New in this version:
4 walls, collision working properly.
Door teleportation phenomena
Objects referencing each other and their code

I need to work on the camera so that the relative angle and position to the player stay the same when teleporting between doors.  I also need to make more of these rooms with more doors and get the doors to actually select their partners now that I understand the architecture better.