Tidied up the transforms of the room objects to make testing easier, fixed teleportation to use transform.InverseTransformDirection and InverseTransformPoint in order to calculate relativity.
After spending a multitude of days trying to figure out why I couldn't understand the matrix math, I have realized that the issue was simply that the players rotation was LOCKED through unity on the X and Z axis.  Boxes unchecked, works beautifully now.

The next step will be to get doors to render real-time textures of the camera projection relative to their partners.